+++ packages/quick_start_browser/vite.config.ts
@@ -9,16 +9,16 @@ export default defineConfig({
     },
     // Used for local builds to force refresh links to the local version
     optimizeDeps: {
-        include: ['azure-kusto-data', 'azure-kusto-ingest'],
-        force:true
-      },
+        include: ["azure-kusto-data", "azure-kusto-ingest"],
+        force: true,
+    },
     build: {
         rollupOptions: {
-          external: ['azure-kusto-data', 'azure-kusto-ingest']
-        }
-      },
+            external: ["azure-kusto-data", "azure-kusto-ingest"],
+        },
+    },
     server: {
         // The port must match the configured redirectUri in the Azure portal app
-        port:   3000
-    }
+        port: 3000,
+    },
 });