+++ CHANGELOG.md
@@ -9,11 +9,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Changed
 
-- IngestClient returns Promise<IngestionResult> instead of Promise<QueueSendMessageResponse>
+-   IngestClient returns Promise<IngestionResult> instead of Promise<QueueSendMessageResponse>
 
 ### Added
 
-- Support table status reporting like explained here https://learn.microsoft.com/en-us/azure/data-explorer/kusto/api/netfx/kusto-ingest-client-status
+-   Support table status reporting like explained here https://learn.microsoft.com/en-us/azure/data-explorer/kusto/api/netfx/kusto-ingest-client-status
+-   When working with storage resources, use success/failure statistics and retry mechanism to improve ingestion stability
 
 ## [5.2.3] - 2023-11-07